tweaked: keep files if you pass in --keep argument, that way if there's a new file you can use the srpm to update our package.

+++ version_check.sh
@@ -1,4 +1,10 @@
 #!/bin/bash
+
+if [ "$1" == "--keep" ]
+then
+  KEEP=true
+fi
+
 DATE="$(date +%s)"
 echo -e "Checking for versions..."
 for x in anaconda firstboot kde-settings korora-backgrounds korora-release shared-mime-info
@@ -12,16 +18,22 @@ do
   then x="generic-release"
   fi
   echo $x
-  mkdir -p tmp-${DATE}/$x
-  cd tmp-${DATE}/$x
+  mkdir -p tmp-${DATE}/$x/{stable,testing}
+  cd tmp-${DATE}/$x/stable
   yumdownloader --source --disablerepo=korora $x >/dev/null 2>&1
   f_version=$(ls ${x}* |awk -F "${x}-" '{print $2}' |awk -F ".fc" '{print $1}')
-  rm -f *
+  cd - >/dev/null 2&>1
+  cd tmp-${DATE}/$x/testing
   yumdownloader --source --enablerepo=updates-testing --disablerepo=korora $x >/dev/null 2>&1
   f_version_testing=$(ls ${x}* |awk -F "${x}-" '{print $2}' |awk -F ".fc" '{print $1}')
-  cd -
-  echo "Korora Version = ${k_version}-${k_release}"
-  echo -e "Fedora Version = $f_version"
-  echo -e "Fedora Version (testing) = $f_version_testing\n"
+  cd - >/dev/null 2&>1
+  echo "${k_version}-${k_release} - Korora Version"
+  echo -e "$f_version - Fedora Version (stable)"
+  echo -e "$f_version_testing - Fedora Version (testing)\n"
 done
-rm -Rf tmp-${DATE}
+if [ ${KEEP} ]
+then
+  echo "Files kept at tmp-${DATE}/"
+else
+  rm -Rf tmp-${DATE}
+fi
